Done LAB3

+++ LAB1.5/HinhCN.cs
@@ -8,9 +8,9 @@ namespace LAB1._5
 {
     internal class HinhCN : Hinh
     {
-        public float ChieuDai { get; set; }
-        public float ChieuRong { get; set; }
-        public override float TinhChuVi() => 2 * (ChieuDai + ChieuRong);
-        public override float TinhDienTich() => ChieuDai * ChieuRong;
+        public double ChieuDai { get; set; }
+        public double ChieuRong { get; set; }
+        public override double TinhChuVi() => 2 * (ChieuDai + ChieuRong);
+        public override double TinhDienTich() => ChieuDai * ChieuRong;
     }
 }
